@@ -1,6 +1,6 @@
 module github.com/nginxinc/nginx-asg-sync
 
-go 1.21.3
+go 1.22.5
 
 require (
 	github.com/Azure/azure-sdk-for-go v68.0.0+incompatible
