@@ -8,18 +8,18 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
-// commonConfig stores the configuration parameters common to all providers
+// commonConfig stores the configuration parameters common to all providers.
 type commonConfig struct {
-	APIEndpoint           string        `yaml:"api_endpoint"`
-	CloudProvider         string        `yaml:"cloud_provider"`
-	SyncIntervalInSeconds time.Duration `yaml:"sync_interval_in_seconds"`
+	APIEndpoint   string        `yaml:"api_endpoint"`
+	CloudProvider string        `yaml:"cloud_provider"`
+	SyncInterval  time.Duration `yaml:"sync_interval_in_seconds"`
 }
 
 func parseCommonConfig(data []byte) (*commonConfig, error) {
 	cfg := &commonConfig{}
 	err := yaml.Unmarshal(data, cfg)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("couldn't unmarshal common config: %w", err)
 	}
 
 	err = validateCommonConfig(cfg)
@@ -35,7 +35,7 @@ func validateCommonConfig(cfg *commonConfig) error {
 		return fmt.Errorf(errorMsgFormat, "api_endpoint")
 	}
 
-	if cfg.SyncIntervalInSeconds == 0 {
+	if cfg.SyncInterval == 0 {
 		return errors.New(intervalErrorMsg)
 	}
 
@@ -50,7 +50,7 @@ func validateCommonConfig(cfg *commonConfig) error {
 	return nil
 }
 
-// Upstream is the cloud agnostic representation of an Upstream (eg, common fields for every cloud provider)
+// Upstream is the cloud agnostic representation of an Upstream (eg, common fields for every cloud provider).
 type Upstream struct {
 	MaxConns     *int
 	MaxFails     *int
